+++ bootstrap/golang/prep-new-repo.sh
@@ -1,5 +1,3 @@
-# TODO DEBUGGING
-set -x
 set -euo pipefail
 
 # =============================================================================
@@ -12,6 +10,10 @@ GO_MOD_FILENAME="go.mod"
 GO_MOD_MODULE_KEYWORD="module "  # The key we'll look for when replacing the module name in go.mod
 
 KURTOSIS_LIB_MODULE="github.com/kurtosis-tech/kurtosis-libs/golang"
+
+# Frustratingly, there's no way to say "do in-place replacement" in sed that's compatible on both Mac and Linux
+# Instead, we add this suffix and delete the backup files after
+SED_INPLACE_FILE_SUFFIX=".sedreplace"
 
 # =============================================================================
 #                             Arg-Parsing & Validation
@@ -64,12 +66,12 @@ fi
 
 # Replace module names in code (we need the "-i '' " argument because Mac sed requires it)
 existing_module_name="$(grep "module" "${go_mod_filepath}" | awk '{print $2}')"
-if ! sed -i '' "s,${existing_module_name},${new_module_name},g" ${go_mod_filepath}; then
+if ! sed -i"${SED_INPLACE_FILE_SUFFIX}" "s,${existing_module_name},${new_module_name},g" ${go_mod_filepath}; then
     echo "Error: Could not replace Go module name in mod file '${go_mod_filepath}'" >&2
     exit 1
 fi
 # We search for old_module_name/testsuite because we don't want the old_module_name/lib entries to get renamed
-if ! sed -i '' "s,${existing_module_name}/${TESTSUITE_IMPL_DIRNAME},${new_module_name}/${TESTSUITE_IMPL_DIRNAME},g" $(find "${output_dirpath}" -type f); then
+if ! sed -i"${SED_INPLACE_FILE_SUFFIX}" "s,${existing_module_name}/${TESTSUITE_IMPL_DIRNAME},${new_module_name}/${TESTSUITE_IMPL_DIRNAME},g" $(find "${output_dirpath}" -type f); then
     echo "Error: Could not replace Go module name in code files" >&2
     exit 1
 fi
@@ -78,5 +80,11 @@ fi
 # Lastly, depend on the actual Kurtosis library
 if ! ( cd "${output_dirpath}" && go get "${KURTOSIS_LIB_MODULE}" ); then
     echo "Error: Failed to pull Kurtosis Go lib dependency '${KURTOSIS_LIB_MODULE}'" >&2
+    exit 1
+fi
+
+# NOTE: Leave this as the last command in the file!! It removes all the backup files created by our in-place sed (see above for why this is necessary)
+if ! find "${output_dirpath}" -name "*${SED_INPLACE_FILE_SUFFIX}" -delete; then
+    echo "Error: Failed to remove the backup files suffixed with '${SED_INPLACE_FILE_SUFFIX}' that we created doing in-place string replacement with sed" >&2
     exit 1
 fi